requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.22
django-allauth==0.56.1
django-crispy-forms==1.14.0
greenlet==3.0.0
gunicorn==21.2.0
oauthlib==3.2.2
psycopg2==2.9.7
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
urllib3==1.26.16
